Fix build warnings with MSVC

+++ test/BaseDecoderTest.h
@@ -37,7 +37,7 @@ class BaseDecoderTest {
   ISVCDecoder* decoder_;
 
  private:
-  void DecodeFrame (const uint8_t* src, int sliceSize, Callback* cbk);
+  void DecodeFrame (const uint8_t* src, size_t sliceSize, Callback* cbk);
 
   std::ifstream file_;
   BufferedData buf_;
